Add react-router routes and Supabase protected route
