+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "lucide-react": "^0.542.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.1",
+    "@supabase/supabase-js": "^2.45.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
