+++ src/pages/Login.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Login() {
+  return (
+    <div className="p-4">Login Page</div>
+  );
+}
