+++ src/pages/Welcome.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Welcome() {
+  return (
+    <div className="p-4">Welcome Page</div>
+  );
+}
