+++ src/types/external.d.ts
@@ -1,0 +1,6 @@
+declare module 'react-router-dom';
+
+declare module '@supabase/supabase-js' {
+  export type Session = unknown;
+  export function createClient(url: string, key: string): unknown;
+}
